Complete Delete User Controller
Commit Create User

+++ Back/controllers/userCn.js
@@ -1,64 +1,67 @@
 import catchAsync from "../utils/catchAsync.js";
 import Users from "../models/Users.js";
-import { mailerSend , emailParams } from "../middlewares/sendEmail.js";
+import { mailerSend, emailParams } from "../middlewares/sendEmail.js";
 import jwt from "jsonwebtoken";
 import bcryptjs from "bcryptjs";
 
-export const createUser = catchAsync(async (req,res,next)=>{
-    try{
-        const {email,id,password,username,name,role,...others} = req.body;
-        const res = await mailerSend.send(emailParams);
-        
-        // Start Check Email Result
-        // **
-        // End Of Check Email Result
+// export const createUser = catchAsync(async (req,res,next)=>{
+//     try{
+//         const {email,id,password,username,name,role,...others} = req.body;
+//         const res = await mailerSend.send(emailParams);
 
-        const hashPass = bcryptjs.hashSync(password);
+//         // Start Check Email Result
+//         // **
+//         // End Of Check Email Result
 
-        const token = jwt.sign({username,id,role,password:hashPass},process.env.JWT_SECRET);
-        
-        const user = await Users.create({email,password,username,name,...others});
+//         const hashPass = bcryptjs.hashSync(password);
 
-        return res.status(201).json({
-            status:"success",
-            message: "User Created Successfully",
-            token,
-        });
-    }catch(err){
-        return res.status(400).json({
-            status: "failed",
-            message: err.message,
-        });
-    };
-});
+//         const token = jwt.sign({username,id,role,password:hashPass},process.env.JWT_SECRET);
 
-export const deleteUser = catchAsync(async(req,res,next)=>{
-    try{
+//         const user = await Users.create({email,password,username,name,...others});
+
+//         return res.status(201).json({
+//             status:"success",
+//             message: "User Created Successfully",
+//             token,
+//         });
+//     }catch(err){
+//         return res.status(400).json({
+//             status: "failed",
+//             message: err.message,
+//         });
+//     };
+// });
+
+export const deleteUser = catchAsync(async (req, res, next) => {
+    const { id, role } = jwt.verify(req.headers.authorization.split(" ")[1], process.env.JWT_SECRET);
+    if (id === req.params.id || role === 'admin' || role === 'superAdmin') {
         await Users.findByIdAndDelete(req.params.id);
         return res.status(200).json({
             status: "success",
             message: `User With Id ${req.params.id} Deleted Successfully`,
         });
-    }catch(err){
+    } else {
         return res.status(400).json({
             status: "failed",
-            message: err.message,
+            message: "You do not have permission to access this action",
         });
     }
+
+
 });
 
-export const getUser = catchAsync(async(req,res,next)=>{
-    try{
+export const getUser = catchAsync(async (req, res, next) => {
+    try {
         // decode id from token
         const user = await Users.findById(req.id).select('-_v,-password');
         return res.status(200).json({
-            status:"success",
+            status: "success",
             message: "User Founded Successfully",
             data: user
         });
-    }catch(err){
+    } catch (err) {
         return res.status(400).json({
-            status:"failed",
+            status: "failed",
             message: err.message,
         });
     };
